add helpers foreign trade hs codes
